refactor: move product components to modules/products directory
