+++ app/(app)/(home)/[category]/[subcategory]/page.tsx
@@ -1,5 +1,5 @@
 import type { SearchParams } from "nuqs/server";
-import { ProductListView } from "@/components/products/ProductListView";
+import ProductListView from "@/modules/products/views/ProductListView";
 import { loadProductFilters } from "@/lib/search-params/product-filters.server";
 import { getQueryClient, trpc } from "@/trpc/server";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
