+++ app/(app)/(home)/page.tsx
@@ -2,7 +2,7 @@ import type { SearchParams } from "nuqs/server";
 import { loadProductFilters } from "@/lib/search-params/product-filters.server";
 import { getQueryClient, trpc } from "@/trpc/server";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
-import { ProductListView } from "@/components/products/ProductListView";
+import ProductListView from "@/modules/products/views/ProductListView";
 import { DEFAULT_LIMIT } from "@/constants";
 
 interface Props {
